+++ CHANGELOG.md
@@ -4,11 +4,9 @@ All notable changes to this project will be documented in this file.
 See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+- Experimental support for token-based authentication in the client and functions for token management.
 
 ## [8.15.0]
-
-### Added
-- Experimental support for token-based authentication in the client and functions for token management.
 - managed load balancer: support for redirect rule HTTP status
 
 ## [8.14.0]