+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+
+### Added
 - Experimental support for token-based authentication in the client and functions for token management.
 
 ## [8.15.0]