+++ CHANGELOG.md
@@ -9,6 +9,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 - Experimental support for token-based authentication in the client and functions for token management.
 
 ## [8.15.0]
+
+### Added
 - managed load balancer: support for redirect rule HTTP status
 
 ## [8.14.0]