+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [8.16.1]
+
 ### Added
 - kubernetes: add `Deprecated` field to `KubernetesPlan`
 
@@ -566,7 +568,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.16.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.16.1...HEAD
+[8.16.1]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.16.0...v8.16.1
 [8.16.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.15.0...v8.16.0
 [8.15.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.14.0...v8.15.0
 [8.14.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.13.0...v8.14.0